Landscape layouts added.
The main motivation was for Means, where the default layout was leaving too little vertical space for the edit views. So I moved the Analyze button up to the same line as the hypothesis relation fields. The other activities were then modified for a consistent look.
Also set maxEms for the text fields in Mean so they don't crowd out the Analyze button.